[💄style]: interface를 type.ts에 분리하여 작성
